Implemented the deletion functionality.

+++ app/Http/Controllers/FileController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\DestroyFilesRequest;
 use App\Http\Requests\StoreFileRequest;
 use App\Http\Requests\StoreFolderRequest;
 use App\Http\Resources\FileResource;
+use App\Http\Responses\JsonResponse;
 use App\Models\File;
 use Exception;
 use Illuminate\Http\Request;
@@ -13,6 +14,7 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Log;
 use Inertia\Inertia;
+use Symfony\Component\HttpFoundation\Response as ResponseAlias;
 
 class FileController extends Controller
 {
@@ -131,22 +133,42 @@ class FileController extends Controller
 
     public function destroy(DestroyFilesRequest $request)
     {
-        $data = $request->validated();
-        $parent = $request->parent ?? File::getDefaultRoot(Auth::id());
+        //        dd($request->all());
+        DB::beginTransaction();
+        try {
+            $data = $request->validated();
+            $parent = $request->parent ?? File::getDefaultRoot(Auth::id());
 
-        if ($data['all']) {
-            $children = $parent->children;
+            if ($data['all']) {
+                $children = $parent->children;
 
-            foreach ($children as $child) {
-                $child->delete();
+                foreach ($children as $child) {
+                    $child->delete();
+                }
+            } else {
+                foreach ($data['ids'] ?? [] as $id) {
+                    $file = File::find($id); // use find instead of findOrFail
+                    if ($file) {
+                        $file->delete();
+                    } else {
+                        Log::info("File with id {$id} not found");
+                    }
+                }
             }
-        } else {
-            foreach ($data['ids'] ?? [] as $id) {
-                $file = File::find($id);
-                $file->delete();
-            }
+            DB::commit();
+
+            return response()->json([
+                'status' => JsonResponse::SUCCESS,
+                'message' => 'Files deleted successfully',
+            ], ResponseAlias::HTTP_OK);
+        } catch (Exception $e) {
+            DB::rollBack();
+            Log::error('Failed to delete files', $e->getTrace());
+
+            return response()->json([
+                'status' => JsonResponse::SUCCESS,
+                'message' => 'Files deleted successfully',
+            ], ResponseAlias::HTTP_BAD_REQUEST);
         }
-
-        return to_route('myFiles', ['folder' => $parent->path]);
     }
 }
